tests: reorganize into deploytool package
