+++ tests/deploytool/__init__.py
@@ -1,0 +1,2 @@
+from . import utils
+from . import models
